feat: extract mango cards into separate feature
