+++ app/features/instruction-program-mango/index.ts
@@ -1,0 +1,1 @@
+export { MangoDetailsCard } from './ui/MangoDetailsCard';
