+++ app/features/instruction-program-mango/ui/AddSpotMarketDetailsCard.tsx
@@ -1,9 +1,8 @@
-import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
+import { InstructionCard } from '@components/instruction/InstructionCard';
 
 import { BaseTable } from '@/app/shared/ui/Table';
-
-import { InstructionCard } from '../InstructionCard';
 import { AddSpotMarket, spotMarketFromIndex } from '@explorer/decoder-mango';
+import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
 
 export function AddSpotMarketDetailsCard(props: {
     ix: TransactionInstruction;
