+++ app/features/instruction-program-mango/ui/PlaceSpotOrderDetailsCard.tsx
@@ -1,56 +1,62 @@
 import { Address } from '@components/common/Address';
+import { InstructionCard } from '@components/instruction/InstructionCard';
+
+import { BaseTable } from '@/app/shared/ui/Table';
+import {
+    getSpotMarketFromInstruction,
+    getSpotMarketFromSpotMarketConfig,
+    OrderLotDetails,
+    PlaceSpotOrder,
+} from '@explorer/decoder-mango';
 import { useCluster } from '@providers/cluster';
 import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
 import BN from 'bn.js';
 import { useEffect, useState } from 'react';
 
-import { BaseTable } from '@/app/shared/ui/Table';
-
-import { InstructionCard } from '../InstructionCard';
-import {
-    getPerpMarketFromInstruction,
-    getPerpMarketFromPerpMarketConfig,
-    OrderLotDetails,
-    PlacePerpOrder,
-} from '@explorer/decoder-mango';
-
-export function PlacePerpOrderDetailsCard(props: {
+export function PlaceSpotOrderDetailsCard(props: {
     ix: TransactionInstruction;
     index: number;
     result: SignatureResult;
-    info: PlacePerpOrder;
+    info: PlaceSpotOrder;
     innerCards?: JSX.Element[];
     childIndex?: number;
 }) {
     const { ix, index, result, info, innerCards, childIndex } = props;
     const mangoAccount = ix.keys[1];
-    const perpMarketAccountMeta = ix.keys[4];
-    const mangoPerpMarketConfig = getPerpMarketFromInstruction(ix, perpMarketAccountMeta);
+    const spotMarketAccountMeta = ix.keys[5];
+    const mangoSpotMarketConfig = getSpotMarketFromInstruction(ix, spotMarketAccountMeta);
 
     const cluster = useCluster();
     const [orderLotDetails, setOrderLotDetails] = useState<OrderLotDetails | null>(null);
     useEffect(() => {
         async function getOrderLotDetails() {
-            if (mangoPerpMarketConfig === undefined) {
+            if (mangoSpotMarketConfig === undefined) {
                 return;
             }
-            const mangoPerpMarket = await getPerpMarketFromPerpMarketConfig(cluster.url, mangoPerpMarketConfig);
-            const maxBaseQuantity = mangoPerpMarket.baseLotsToNumber(new BN(info.quantity.toString()));
-            const limitPrice = mangoPerpMarket.priceLotsToNumber(new BN(info.price.toString()));
+            const mangoSpotMarket = await getSpotMarketFromSpotMarketConfig(
+                ix.programId,
+                cluster.url,
+                mangoSpotMarketConfig,
+            );
+            if (mangoSpotMarket === undefined) {
+                return;
+            }
+            const maxBaseQuantity = mangoSpotMarket.baseSizeLotsToNumber(new BN(info.maxBaseQuantity.toString()));
+            const limitPrice = mangoSpotMarket.priceLotsToNumber(new BN(info.limitPrice.toString()));
             setOrderLotDetails({
                 price: limitPrice,
                 size: maxBaseQuantity,
             } as OrderLotDetails);
         }
         getOrderLotDetails();
-    }, [cluster.url, info.quantity, info.price, mangoPerpMarketConfig]);
+    }, [cluster.url, info.maxBaseQuantity, info.limitPrice, ix.programId, mangoSpotMarketConfig]);
 
     return (
         <InstructionCard
             ix={ix}
             index={index}
             result={result}
-            title="Mango Program: PlacePerpOrder"
+            title="Mango Program: PlaceSpotOrder"
             innerCards={innerCards}
             childIndex={childIndex}
         >
@@ -62,53 +68,51 @@ export function PlacePerpOrderDetailsCard(props: {
                 </BaseTable.Cell>
             </BaseTable.Row>
 
-            {mangoPerpMarketConfig !== undefined && (
+            {mangoSpotMarketConfig !== undefined && (
                 <BaseTable.Row>
-                    <BaseTable.Cell>Perp market</BaseTable.Cell>
-                    <BaseTable.Cell className="text-right">{mangoPerpMarketConfig.name}</BaseTable.Cell>
+                    <BaseTable.Cell>Spot market</BaseTable.Cell>
+                    <BaseTable.Cell className="text-right">{mangoSpotMarketConfig.name}</BaseTable.Cell>
                 </BaseTable.Row>
             )}
 
             <BaseTable.Row>
-                <BaseTable.Cell>Perp market address</BaseTable.Cell>
+                <BaseTable.Cell>Spot market address</BaseTable.Cell>
                 <BaseTable.Cell>
-                    <Address pubkey={perpMarketAccountMeta.pubkey} alignRight link />
+                    <Address pubkey={spotMarketAccountMeta.pubkey} alignRight link />
                 </BaseTable.Cell>
             </BaseTable.Row>
-
-            {info.clientOrderId !== '0' && (
-                <BaseTable.Row>
-                    <BaseTable.Cell>Client order Id</BaseTable.Cell>
-                    <BaseTable.Cell className="text-right">{info.clientOrderId}</BaseTable.Cell>
-                </BaseTable.Row>
-            )}
 
             <BaseTable.Row>
                 <BaseTable.Cell>Order type</BaseTable.Cell>
                 <BaseTable.Cell className="text-right">{info.orderType}</BaseTable.Cell>
             </BaseTable.Row>
+
+            {info.clientId !== '0' && (
+                <BaseTable.Row>
+                    <BaseTable.Cell>Client Id</BaseTable.Cell>
+                    <BaseTable.Cell className="text-right">{info.clientId}</BaseTable.Cell>
+                </BaseTable.Row>
+            )}
+
             <BaseTable.Row>
-                <BaseTable.Cell>side</BaseTable.Cell>
+                <BaseTable.Cell>Side</BaseTable.Cell>
                 <BaseTable.Cell className="text-right">{info.side}</BaseTable.Cell>
             </BaseTable.Row>
 
             {orderLotDetails !== null && (
                 <BaseTable.Row>
-                    <BaseTable.Cell>price</BaseTable.Cell>
+                    <BaseTable.Cell>Limit price</BaseTable.Cell>
+                    {/* todo fix price */}
                     <BaseTable.Cell className="text-right">{orderLotDetails?.price} USDC</BaseTable.Cell>
                 </BaseTable.Row>
             )}
 
             {orderLotDetails !== null && (
                 <BaseTable.Row>
-                    <BaseTable.Cell>quantity</BaseTable.Cell>
+                    <BaseTable.Cell>Size</BaseTable.Cell>
                     <BaseTable.Cell className="text-right">{orderLotDetails?.size}</BaseTable.Cell>
                 </BaseTable.Row>
             )}
-            <BaseTable.Row>
-                <BaseTable.Cell>Reduce only</BaseTable.Cell>
-                <BaseTable.Cell className="text-right">{info.reduceOnly}</BaseTable.Cell>
-            </BaseTable.Row>
         </InstructionCard>
     );
 }
